requirements.txt
psycopg2-binary==2.9.9
flask==2.3.3
flask-cors==4.0.0 